First iteration of making a formal render-vega-lite utility module for validating and materializing the vega-lite-like json that is used to configure a raster layer that is eventually materialized into a vega to pass to the render_vega server endpoint.
