+++ src/mixins/render-vega-lite/Definitions/Encoding/Scale/Continuous/LogScale.js
@@ -1,0 +1,13 @@
+import ContinuousScale from "./ContinuousScale"
+import ScaleType from "../Enums/ScaleType"
+
+export default class LogScale extends ContinuousScale {
+  /**
+   *
+   * @param {Object} scale_definition_object
+   * @param {ParentInfo} parent_info
+   */
+  constructor(scale_definition_object, parent_info) {
+    super(scale_definition_object, ScaleType.kLog, parent_info)
+  }
+}
