+++ src/mixins/render-vega-lite/PropDescriptor/CommonChannelDescriptors.js
@@ -1,0 +1,10 @@
+export { default as ColorChannelDescriptor } from "./CommonChannels/ColorChannelDescriptor"
+export { default as GeographicChannelDescriptor } from "./CommonChannels/GeographicChannelDescriptor"
+export { default as OpacityChannelDescriptor } from "./CommonChannels/OpacityChannelDescriptor"
+export { default as PositionChannelDescriptor } from "./CommonChannels/PositionChannelDescriptor"
+export { default as SizeChannelDescriptor } from "./CommonChannels/SizeChannelDescriptor"
+export {
+  default as PropDescriptor,
+  PropLocation,
+  MeasurementType
+} from "./PropDescriptor"
